balek

+++ models/user.js
@@ -31,7 +31,7 @@ module.exports = (sequelize, DataTypes) => {
       location: DataTypes.JSON,
       phone_number: DataTypes.STRING,
       address: DataTypes.STRING,
-      otp_code: DataTypes.INTEGER,
+      otp_code: DataTypes.DECIMAL,
       status_verified: DataTypes.BOOLEAN
     },
     {
